[Rust SDK][refactor](instance, wasi): Moved wasi related types into wasi module.
